Smooth FPS changes for both drm and X11. Do not use float as frame time for GLSL background as the precision goes down quite quickly during runtime. Instead calculate 0..1 cycle time (0 begin - 1 finished) from pre-specified amount of frames needed for shader.

+++ src/render/render.c
@@ -74,7 +74,7 @@ wlc_render_clear(struct wlc_render *render)
 }
 
 void
-wlc_render_time(struct wlc_render *render, float time)
+wlc_render_time(struct wlc_render *render, uint32_t time)
 {
    assert(render);
    render->api.time(render->render, time);
